[DDS-120] update constants

+++ x/did/types/denom.go
@@ -2,5 +2,5 @@ package types
 
 var (
 	BaseMinimalDenom = "ncheq"
-	BaseFactor       = 1e9
+	BaseFactor       = int64(1e9)
 )
